requirements.txt
numpy
numba
av
Pillow